requirement.txt
streamlit
pandas
matplotlib
seaborn
langchain-experimental
langchain
langchain-groq
python-dotenv